Add files via upload
This guide aims to install GeoKey on a cloud server running on a Ubuntu machine, specifically version 18.04 that will be supported until 2028. 

It explains how to 
    • install pre-requisites & Python packages,
    • setup the database,
    • set up virtual environment,
    • set up GeoKey and 
    • configure the system files and make necessary modifications to make GeoKey run smoothly.

This documentation aims to simplify the processes as much as possible to make it suitable for everyone including non-technical people not familiar with Python / Linux environment but keen on running GeoKey on their own cloud servers.